Bug 6725 - Added a nightly read me and updated the build scripts to use the correct read me file.
git-svn-id: http://publicsvn.songbirdnest.com/client/trunk@7800 5aad4ff5-4b58-4159-ba3f-0e6328df4cad

+++ installer/macosx/make-installer.sh
@@ -1,14 +1,20 @@
 #!/bin/sh
 
+if [ $# < 2 ]; then
+  echo usage: make-installer architecture readme
+  exit 1
+fi
+
 DEPTH=../../..
 CURRENT_DIR=`pwd`
 CURRENT_DATE=`date +%Y%m%d`
 ARCH="$1"
+README_FILE="$2"
 
 rm -rf ${DEPTH}/compiled/_built_installer
 mkdir ${DEPTH}/compiled/_built_installer
 
-cp ${DEPTH}/compiled/dist/Songbird.app/Contents/Resources/README.txt ${DEPTH}/compiled/dist/README.txt
+cp ${DEPTH}/compiled/dist/Songbird.app/Contents/Resources/${README_FILE} ${DEPTH}/compiled/dist/README.txt
 ln -s /Applications/ ${DEPTH}/compiled/dist/Applications
 
 ${DEPTH}/installer/macosx/make-diskimage \
